+++ src/comet/__init__.py
@@ -1,5 +1,5 @@
 """Constant-Pressure simulations using Metropolis Algorithm package scaffolding."""
 
-__version__ = "0.3.0"
+__version__ = "0.3.1"
 
 __all__ = ["__version__"]